% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-widgets}
\alias{test-widgets}
\title{Test case: widgets}
\description{
Test case: widgets
}
\examples{
if (requireNamespace("htmltools", quietly = TRUE)) {
  htmlwidgets::setWidgetIdSeed(42)
}

library(leaflet)
leaflet() \%>\%
  addTiles() \%>\%
  addMarkers(
    lng=174.768,
    lat=-36.852,
    popup="The birthplace of R"
  )

library(heatmaply)
heatmaply(mtcars, k_row = 3, k_col = 2)

library(plotly)
plot_ly(midwest,
  x = ~percollege,
  color = ~state,
  type = "box"
)

library(DiagrammeR)
grViz("
  digraph {
  layout = twopi
  node [shape = circle]
  A -> {B C D}
}")
}
\seealso{
Other tests: \code{\link{test-dont}},
  \code{\link{test-figures}}, \code{\link{test-links}},
  \code{\link{test-lists}}, \code{\link{test-params}}
}
\keyword{internal}
